requirements.txt
dash==2.4.1
dash-bootstrap-components==1.1.0
user-agents==2.2.0
gunicorn==20.0.4
requests==2.25.0
dash-uploader==0.6.0
dash-mantine-components==0.10.2
git+https://github.com/pyronear/pyro-storage.git#subdirectory=client
python-dotenv>=0.15.0
python-dateutil==2.8.2
